requirements.txt
praw
supabase
discord.py
openai>=1.0.0
